poprawki

+++ src/main/java/com/crud/library/repository/BookDao.java
@@ -5,18 +5,20 @@ import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import java.math.BigDecimal;
+import java.util.List;
 
 @Repository
 public class BookDao {
     @PersistenceContext
     private EntityManager em;
 
-    public Long getAllAvialableCopies(Long bookId) {
+    public BigDecimal getAllAvialableCopies(Long bookId) {
         String SearchQuery = "SELECT count(*) FROM LIBRARY_CRUD.COPIES c\n" +
                 "JOIN LIBRARY_CRUD.BOOKS B\n" +
                 "ON B.ID = C.BOOK_ID\n" +
                 "WHERE C.STATUS='FREE' AND B.ID= " + bookId;
-        return (Long) em.createNativeQuery(SearchQuery).getSingleResult();
+        return (BigDecimal) em.createNativeQuery(SearchQuery).getSingleResult();
     }
 
     public RentalDto returnBorrowedBook(Long copiesId) {
